+++ board_ws/src/board_ros/src/function.cpp
@@ -28,35 +28,35 @@ Target targetArray[7]; // 储存靶标信息的数组
 
 
 /* ============ 全局变量（静态，外部不可见） ============ */
-static float x_[4] = {0};        // [x, y, vx, vy]
-static float P_[4][4] = {{0}};   // 协方差矩阵
+static double x_[4] = {0};        // [x, y, vx, vy]
+static double P_[4][4] = {{0}};   // 协方差矩阵
 static bool  firstCall_ = true;  // 自动用第一次观测做初始化
 /* ======================================================= */
 
 /* 4x4 矩阵乘法：C = A*B */
-static void mul44(const float A[4][4], const float B[4][4], float C[4][4])
+static void mul44(const double A[4][4], const double B[4][4], double C[4][4])
 {
     for (int i = 0; i < 4; ++i)
         for (int j = 0; j < 4; ++j)
         {
-            float s = 0;
+            double s = 0;
             for (int k = 0; k < 4; ++k) s += A[i][k] * B[k][j];
             C[i][j] = s;
         }
 }
 
 /* 4x4 矩阵加常数对角：A += diag(v) */
-static void addDiag44(float A[4][4], float v)
+static void addDiag44(double A[4][4], double v)
 {
     for (int i = 0; i < 4; ++i) A[i][i] += v;
 }
 
 /* 求 2x2 逆矩阵 */
-static bool inv22(const float M[2][2], float Minv[2][2])
+static bool inv22(const double M[2][2], double Minv[2][2])
 {
-    float det = M[0][0]*M[1][1] - M[0][1]*M[1][0];
+    double det = M[0][0]*M[1][1] - M[0][1]*M[1][0];
     if (std::fabs(det) < 1e-6f) return false;
-    float invDet = 1.0f / det;
+    double invDet = 1.0f / det;
     Minv[0][0] =  M[1][1] * invDet;
     Minv[0][1] = -M[0][1] * invDet;
     Minv[1][0] = -M[1][0] * invDet;
@@ -67,11 +67,11 @@ static bool inv22(const float M[2][2], float Minv[2][2])
 // 卡尔曼滤波
 /*
 使用示例：
-float x, y, vx, vy;
+double x, y, vx, vy;
 KalmanUpdate(obsX, obsY, x, y, vx, vy);
 printf("pos=(%.2f,%.2f)  vel=(%.2f,%.2f)\n", x, y, vx, vy);
 */
-void KalmanUpdate(float obsX, float obsY, float& outX, float& outY,float& outVx, float& outVy)
+void KalmanUpdate(double obsX, double obsY, double& outX, double& outY,double& outVx, double& outVy)
 {
     /* 第一次调用：用观测初始化状态，协方差置单位阵 */
     if (firstCall_)
@@ -84,18 +84,18 @@ void KalmanUpdate(float obsX, float obsY, float& outX, float& outY,float& outVx,
     }
 
     /* 1. 预测（恒定速度模型，dt=1） */
-    float F[4][4] = {{1,0,1,0}, {0,1,0,1}, {0,0,1,0}, {0,0,0,1}};
-    float Q[4][4] = {{0}};
+    double F[4][4] = {{1,0,1,0}, {0,1,0,1}, {0,0,1,0}, {0,0,0,1}};
+    double Q[4][4] = {{0}};
 
     //TODO:目前通过速度判断靶标是否为静止靶标或者移动靶，之后需要将速度条件判断变成靶标类别判断，使得KF适用于静止和移动
-    static float qStill = 1e-4f;     // 静止用
-    static float qMove  = 0.1f;      // 运动用
-    float spd = std::fabs(x_[2]) + std::fabs(x_[3]);
-    float qScale = (spd < 0.02f) ? qStill : qMove;   // 速度阈值 0.02 可改
+    static double qStill = 1e-4f;     // 静止用
+    static double qMove  = 0.1f;      // 运动用
+    double spd = std::fabs(x_[2]) + std::fabs(x_[3]);
+    double qScale = (spd < 0.02f) ? qStill : qMove;   // 速度阈值 0.02 可改
     addDiag44(Q, qScale);          // 过程噪声强度可调
 
     /* x = F*x */
-    float tmp[4];
+    double tmp[4];
     for (int i = 0; i < 4; ++i)
     {
         tmp[i] = 0;
@@ -104,30 +104,30 @@ void KalmanUpdate(float obsX, float obsY, float& outX, float& outY,float& outVx,
     for (int i = 0; i < 4; ++i) x_[i] = tmp[i];
 
     /* P = F*P*F' + Q */
-    float FP[4][4];
+    double FP[4][4];
     mul44(F, P_, FP);
-    float FT[4][4];
+    double FT[4][4];
     for (int i = 0; i < 4; ++i)
         for (int j = 0; j < 4; ++j) FT[i][j] = F[j][i];
     mul44(FP, FT, P_);
     addDiag44(P_, 0.1f);   // 即 Q 的对角
 
     /* 2. 更新 */
-    float H[2][4] = {{1,0,0,0}, {0,1,0,0}};
-    float R[2][2] = {{1,0}, {0,1}};   // 观测噪声强度可调
+    double H[2][4] = {{1,0,0,0}, {0,1,0,0}};
+    double R[2][2] = {{1,0}, {0,1}};   // 观测噪声强度可调
 
     /* y = z - H*x */
-    float y[2] = { obsX - x_[0], obsY - x_[1] };
+    double y[2] = { obsX - x_[0], obsY - x_[1] };
 
     /* S = H*P*H' + R */
-    float HP[2][4];
+    double HP[2][4];
     for (int i = 0; i < 2; ++i)
         for (int j = 0; j < 4; ++j)
         {
             HP[i][j] = 0;
             for (int k = 0; k < 4; ++k) HP[i][j] += H[i][k] * P_[k][j];
         }
-    float S[2][2] = {{0}};
+    double S[2][2] = {{0}};
     for (int i = 0; i < 2; ++i)
         for (int j = 0; j < 2; ++j)
         {
@@ -136,9 +136,9 @@ void KalmanUpdate(float obsX, float obsY, float& outX, float& outY,float& outVx,
         }
 
     /* K = P*H'*inv(S) */
-    float Sinv[2][2];
+    double Sinv[2][2];
     if (!inv22(S, Sinv)) return;          // 奇异则跳过更新
-    float K[4][2] = {{0}};
+    double K[4][2] = {{0}};
     for (int i = 0; i < 4; ++i)
         for (int j = 0; j < 2; ++j)
             for (int k = 0; k < 2; ++k)
@@ -149,16 +149,16 @@ void KalmanUpdate(float obsX, float obsY, float& outX, float& outY,float& outVx,
     for (int i = 0; i < 4; ++i) x_[i] += K[i][0]*y[0] + K[i][1]*y[1];
 
     /* P = (I - K*H)*P */
-    float KH[4][4] = {{0}};
+    double KH[4][4] = {{0}};
     for (int i = 0; i < 4; ++i)
         for (int j = 0; j < 4; ++j)
             for (int k = 0; k < 2; ++k)
                 KH[i][j] += K[i][k] * H[k][j];
-    float IKH[4][4];
+    double IKH[4][4];
     for (int i = 0; i < 4; ++i)
         for (int j = 0; j < 4; ++j)
             IKH[i][j] = (i==j) - KH[i][j];
-    float tmpP[4][4];
+    double tmpP[4][4];
     mul44(IKH, P_, tmpP);
     for (int i = 0; i < 4; ++i)
         for (int j = 0; j < 4; ++j) P_[i][j] = tmpP[i][j];
